Back-end/requirements.txt

Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0

psycopg==3.2.10
PyJWT==2.10.1
python-dotenv==1.1.1


